split tests to run them concurrently. Add test for blocking

+++ tests/feature_test.go
@@ -3,6 +3,7 @@
 package tests
 
 import (
+	"errors"
 	"flag"
 	"fmt"
 	"io/ioutil"
@@ -26,10 +27,10 @@ import (
 var opt = godog.Options{
 	Paths:  []string{"."},
 	Output: colors.Colored(os.Stdout),
-	Format: "pretty",
-	// Format:      "progress",
-	// Concurrency: 4,
-	Randomize: time.Now().UTC().UnixNano(),
+	// Format: "pretty",
+	Format:      "progress",
+	Concurrency: 4,
+	Randomize:   time.Now().UTC().UnixNano(),
 }
 
 func init() {
@@ -53,8 +54,12 @@ func TestMain(m *testing.M) {
 }
 
 type feature struct {
-	testID  string
-	testDir string
+	scenarioID string
+	testDir    string
+	featurID   string
+
+	lastOutput string
+	lastErr    error
 
 	cf *cloudformation.CloudFormation
 }
@@ -68,30 +73,27 @@ func (f *feature) fileExists(fname string, content *gherkin.DocString) error {
 		return err
 	}
 
-	c := strings.Replace(content.Content, "%testid%", f.testID, -1)
+	c := strings.Replace(content.Content, "%scenarioid%", f.scenarioID, -1)
+	c = strings.Replace(c, "%featureid%", f.featurID, -1)
+
 	return ioutil.WriteFile(fpath, []byte(c), 0700)
 }
 
 func (f *feature) iSuccessfullyRun(cmd string) error {
-	bin, err := filepath.Abs("../bin/stas")
+	err := f.iRun(cmd)
 	if err != nil {
 		return err
 	}
 
-	c := exec.Command(bin, strings.Split(cmd, " ")...)
-	c.Dir = f.testDir
-
-	out, err := c.CombinedOutput()
-
-	if err != nil {
-		return fmt.Errorf("err: %v, output:\n%s", err, string(out))
+	if f.lastErr != nil {
+		return fmt.Errorf("err: %v, output:\n%s", err, string(f.lastOutput))
 	}
 
 	return nil
 }
 
 func (f *feature) stackShouldHaveStatus(stackName, status string) error {
-	s := strings.Replace(stackName, "%testid%", f.testID, -1)
+	s := strings.Replace(stackName, "%scenarioid%", f.scenarioID, -1)
 	out, err := f.cf.DescribeStacks(&cloudformation.DescribeStacksInput{
 		StackName: aws.String(s),
 	})
@@ -103,6 +105,40 @@ func (f *feature) stackShouldHaveStatus(stackName, status string) error {
 		return fmt.Errorf("stack status is %s", aws.StringValue(out.Stacks[0].StackStatus))
 	}
 
+	return nil
+}
+
+func (f *feature) iModifyFile(fname string, content *gherkin.DocString) error {
+	return f.fileExists(fname, content)
+}
+
+func (f *feature) iRun(cmd string) error {
+	bin, err := filepath.Abs("../bin/stas")
+	if err != nil {
+		return err
+	}
+
+	c := exec.Command(bin, strings.Split(cmd, " ")...)
+	c.Dir = f.testDir
+
+	out, err := c.CombinedOutput()
+	f.lastOutput = string(out)
+	f.lastErr = err
+
+	return nil
+}
+
+func (f *feature) exitCodeShouldNotBeZero() error {
+	if f.lastErr == nil {
+		return errors.New("program returned zero exit code")
+	}
+	return nil
+}
+
+func (f *feature) outputShouldContain(s *gherkin.DocString) error {
+	if !strings.Contains(f.lastOutput, s.Content) {
+		return fmt.Errorf("output doesn't contain searched string. Output:\n%s", f.lastOutput)
+	}
 	return nil
 }
 
@@ -118,15 +154,24 @@ func FeatureContext(s *godog.Suite) {
 	s.Step(`^file "([^"]*)" exists:$`, f.fileExists)
 	s.Step(`^I successfully run "([^"]*)"$`, f.iSuccessfullyRun)
 	s.Step(`^stack "([^"]*)" should have status "([^"]*)"$`, f.stackShouldHaveStatus)
+	s.Step(`^I modify file "([^"]*)":$`, f.iModifyFile)
+	s.Step(`^I run "([^"]*)"$`, f.iRun)
+	s.Step(`^exit code should not be zero$`, f.exitCodeShouldNotBeZero)
+	s.Step(`^output should contain:$`, f.outputShouldContain)
 
 	s.BeforeScenario(func(interface{}) {
-		f.testID = strconv.FormatInt(rand.Int63(), 10)
-		f.testDir = filepath.Join(testDir, "stas_test_"+f.testID)
+		f.scenarioID = strconv.FormatInt(rand.Int63(), 10)
+		f.testDir = filepath.Join(testDir, "stas_test_"+f.scenarioID)
+	})
+	s.AfterScenario(func(interface{}, error) {
+		os.RemoveAll(f.testDir)
 	})
 
-	s.AfterSuite(func() {
-		os.RemoveAll(testDir)
+	s.BeforeFeature(func(*gherkin.Feature) {
+		f.featurID = strconv.FormatInt(rand.Int63(), 10)
+	})
 
+	s.AfterFeature(func(*gherkin.Feature) {
 		stacks, err := cf.DescribeStacks(&cloudformation.DescribeStacksInput{})
 		if err != nil {
 			panic(err)
@@ -134,7 +179,7 @@ func FeatureContext(s *godog.Suite) {
 
 		for _, s := range stacks.Stacks {
 			for _, t := range s.Tags {
-				if aws.StringValue(t.Key) == "STAS_TEST" {
+				if aws.StringValue(t.Key) == "STAS_TEST" && aws.StringValue(t.Value) == f.featurID {
 					cf.DeleteStack(&cloudformation.DeleteStackInput{
 						StackName: s.StackName,
 					})
